pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splitwise-bill-splitter"
version = "0.1.0"
description = "Smart Bill Splitter with AI"
readme = "README.md"
authors = [
    {name = "Vikram Singh", email = "vikramsingh9258@gmail.com"}
]
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "pillow>=9.0.0",
    "openai>=1.0.0",
    "dotenv>=0.19.0",
    "pydantic-ai-slim[openai]>=0.4.3",
    "loguru>=0.7.3",
    "nest-asyncio>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.scripts]
splitwise = "splitwise.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
